feat: 代码重构

+++ app/system/internal/data/model/role_resource.go
@@ -22,7 +22,7 @@ func (t ResourceType) String() string {
 // 角色资源关联表
 type RoleResource struct {
 	ID         uint         `gorm:"primarykey"`
-	RoleID     uint64       `gorm:"index;default:0;not null;comment:角色ID"`                 // 角色ID
+	RoleID     uint         `gorm:"index;default:0;not null;comment:角色ID"`                 // 角色ID
 	ResourceID string       `gorm:"index;default:'';not null;comment:资源ID"`                // 资源ID
 	Type       ResourceType `gorm:"index;type:tinyint(1);default:0;not null;comment:资源类型"` // 资源类型
 }
